wip attempting to write link checker.
Wrote some code in gulpfile to parse markdown links and return links.
I believe it's working.

Realized the "link checker" tried to crawl the website itself.
I wrote a bit of code to start the webserver and run the checked, but it
errors out with errors which should be benign.